gencode: Move to own pkg
