+++ internal/serializers/gencode/gencode-unsafe.schema
