+++ internal/serializers/gencode/gencode-unsafe.schema.gen.go
@@ -1,4 +1,4 @@
-package goserbench
+package gencode
 
 import (
 	"io"
@@ -12,16 +12,16 @@ var (
 	_ = time.Now()
 )
 
-type GencodeA struct {
+type GencodeUnsafeA struct {
 	Name     string
-	BirthDay time.Time
+	BirthDay int64
 	Phone    string
 	Siblings int32
 	Spouse   bool
 	Money    float64
 }
 
-func (d *GencodeA) Size() (s uint64) {
+func (d *GencodeUnsafeA) Size() (s uint64) {
 
 	{
 		l := uint64(len(d.Name))
@@ -67,10 +67,10 @@ func (d *GencodeA) Size() (s uint64) {
 		s++
 
 	}
-	s += 24
+	s += 17
 	return
 }
-func (d *GencodeA) Marshal(buf []byte) ([]byte, error) {
+func (d *GencodeUnsafeA) Marshal(buf []byte) ([]byte, error) {
 	size := d.Size()
 	{
 		if uint64(cap(buf)) >= size {
@@ -101,11 +101,9 @@ func (d *GencodeA) Marshal(buf []byte) ([]byte, error) {
 		i += l
 	}
 	{
-		b, err := d.BirthDay.MarshalBinary()
-		if err != nil {
-			return nil, err
-		}
-		copy(buf[i+0:], b)
+
+		*(*int64)(unsafe.Pointer(&buf[i+0])) = d.BirthDay
+
 	}
 	{
 		l := uint64(len(d.Phone))
@@ -115,15 +113,15 @@ func (d *GencodeA) Marshal(buf []byte) ([]byte, error) {
 			t := uint64(l)
 
 			for t >= 0x80 {
-				buf[i+15] = byte(t) | 0x80
+				buf[i+8] = byte(t) | 0x80
 				t >>= 7
 				i++
 			}
-			buf[i+15] = byte(t)
+			buf[i+8] = byte(t)
 			i++
 
 		}
-		copy(buf[i+15:], d.Phone)
+		copy(buf[i+8:], d.Phone)
 		i += l
 	}
 	{
@@ -136,46 +134,30 @@ func (d *GencodeA) Marshal(buf []byte) ([]byte, error) {
 		}
 
 		for t >= 0x80 {
-			buf[i+15] = byte(t) | 0x80
+			buf[i+8] = byte(t) | 0x80
 			t >>= 7
 			i++
 		}
-		buf[i+15] = byte(t)
+		buf[i+8] = byte(t)
 		i++
 
 	}
 	{
 		if d.Spouse {
-			buf[i+15] = 1
+			buf[i+8] = 1
 		} else {
-			buf[i+15] = 0
+			buf[i+8] = 0
 		}
 	}
 	{
 
-		v := *(*uint64)(unsafe.Pointer(&(d.Money)))
-
-		buf[i+0+16] = byte(v >> 0)
-
-		buf[i+1+16] = byte(v >> 8)
-
-		buf[i+2+16] = byte(v >> 16)
-
-		buf[i+3+16] = byte(v >> 24)
-
-		buf[i+4+16] = byte(v >> 32)
-
-		buf[i+5+16] = byte(v >> 40)
-
-		buf[i+6+16] = byte(v >> 48)
-
-		buf[i+7+16] = byte(v >> 56)
+		*(*float64)(unsafe.Pointer(&buf[i+9])) = d.Money
 
 	}
-	return buf[:i+24], nil
+	return buf[:i+17], nil
 }
 
-func (d *GencodeA) Unmarshal(buf []byte) (uint64, error) {
+func (d *GencodeUnsafeA) Unmarshal(buf []byte) (uint64, error) {
 	i := uint64(0)
 
 	{
@@ -199,7 +181,9 @@ func (d *GencodeA) Unmarshal(buf []byte) (uint64, error) {
 		i += l
 	}
 	{
-		d.BirthDay.UnmarshalBinary(buf[i+0 : i+0+15])
+
+		d.BirthDay = *(*int64)(unsafe.Pointer(&buf[i+0]))
+
 	}
 	{
 		l := uint64(0)
@@ -207,10 +191,10 @@ func (d *GencodeA) Unmarshal(buf []byte) (uint64, error) {
 		{
 
 			bs := uint8(7)
-			t := uint64(buf[i+15] & 0x7F)
-			for buf[i+15]&0x80 == 0x80 {
+			t := uint64(buf[i+8] & 0x7F)
+			for buf[i+8]&0x80 == 0x80 {
 				i++
-				t |= uint64(buf[i+15]&0x7F) << bs
+				t |= uint64(buf[i+8]&0x7F) << bs
 				bs += 7
 			}
 			i++
@@ -218,16 +202,16 @@ func (d *GencodeA) Unmarshal(buf []byte) (uint64, error) {
 			l = t
 
 		}
-		d.Phone = string(buf[i+15 : i+15+l])
+		d.Phone = string(buf[i+8 : i+8+l])
 		i += l
 	}
 	{
 
 		bs := uint8(7)
-		t := uint32(buf[i+15] & 0x7F)
-		for buf[i+15]&0x80 == 0x80 {
+		t := uint32(buf[i+8] & 0x7F)
+		for buf[i+8]&0x80 == 0x80 {
 			i++
-			t |= uint32(buf[i+15]&0x7F) << bs
+			t |= uint32(buf[i+8]&0x7F) << bs
 			bs += 7
 		}
 		i++
@@ -239,13 +223,12 @@ func (d *GencodeA) Unmarshal(buf []byte) (uint64, error) {
 
 	}
 	{
-		d.Spouse = buf[i+15] == 1
+		d.Spouse = buf[i+8] == 1
 	}
 	{
 
-		v := 0 | (uint64(buf[i+0+16]) << 0) | (uint64(buf[i+1+16]) << 8) | (uint64(buf[i+2+16]) << 16) | (uint64(buf[i+3+16]) << 24) | (uint64(buf[i+4+16]) << 32) | (uint64(buf[i+5+16]) << 40) | (uint64(buf[i+6+16]) << 48) | (uint64(buf[i+7+16]) << 56)
-		d.Money = *(*float64)(unsafe.Pointer(&v))
+		d.Money = *(*float64)(unsafe.Pointer(&buf[i+9]))
 
 	}
-	return i + 24, nil
+	return i + 17, nil
 }
